Restarted player and attack
I restarted the character movement and attack.
Added a magic staff that is like a sword and works with the mouse cursor having the character aim left or right with the wand.
Started an inventory system that toggles between different weapons (not equippable yet).
Added a dash to the character with spacebar
I hope to have a bow or maybe change it to be a gun by tomorrow night and it will aim with the mouse in which direction it will shoot
Slime with random movement but still cannot get the player to destroy the slime